WIP: enchilada: cleanup and add earpiece/VoiceCall
The controls should somewhat better reflect the hardware, earpiece also
works now!

There's a bug in ucm where if multiple "SectionDevice" configs specify
the same PCM ID (in this case the headphones and earpiece) they must
have the exact same Values configuration, otherwise it leads to
weirdness and pulseaudio gets sad. As a result the earpiece shows up as
a stereo output device but only the left channel works.

As it stands I think there's an error in the HiFi config, pulseaudio on
startup says "Failed to set verb HiFi", maybe indicating that it's
always using the VoiceCall usecase? This needs to be fixed