codecs: wcd934x: configure additional microphones
The WCD934x codec has 3 microphone interfaces, configure them all.
The first microphone was already configured to be used for a headset.
Do the same for the other 2 microphones. This won't influence devices
with less than 3 microphones because the microphones still have to be
enabled in the HiFi or VoiceCall profiles.